# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CyberCat",
    "description": "Threat-informed automated incident response platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/responses": {
      "get": {
        "tags": [
          "responses"
        ],
        "summary": "List Responses",
        "operationId": "list_responses_v1_responses_get",
        "parameters": [
          {
            "name": "incident_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Incident Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "classification",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Classification"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO 8601 datetime; return actions proposed after this time",
              "title": "Since"
            },
            "description": "ISO 8601 datetime; return actions proposed after this time"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "responses"
        ],
        "summary": "Propose Response",
        "operationId": "propose_response_v1_responses_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionProposeIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionProposed"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/v1/responses/{action_id}/execute": {
      "post": {
        "tags": [
          "responses"
        ],
        "summary": "Execute Response",
        "operationId": "execute_response_v1_responses__action_id__execute_post",
        "parameters": [
          {
            "name": "action_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Action Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionExecuted"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/v1/responses/{action_id}/revert": {
      "post": {
        "tags": [
          "responses"
        ],
        "summary": "Revert Response",
        "operationId": "revert_response_v1_responses__action_id__revert_post",
        "parameters": [
          {
            "name": "action_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Action Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionExecuted"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/lab/assets": {
      "get": {
        "tags": [
          "lab"
        ],
        "summary": "List Lab Assets",
        "operationId": "list_lab_assets_v1_lab_assets_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LabAssetOut"
                  },
                  "type": "array",
                  "title": "Response List Lab Assets V1 Lab Assets Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "lab"
        ],
        "summary": "Register Lab Asset",
        "operationId": "register_lab_asset_v1_lab_assets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabAssetIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabAssetOut"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/lab/assets/{asset_id}": {
      "delete": {
        "tags": [
          "lab"
        ],
        "summary": "Deregister Lab Asset",
        "operationId": "deregister_lab_asset_v1_lab_assets__asset_id__delete",
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Asset Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActionExecuted": {
        "properties": {
          "action": {
            "$ref": "#/components/schemas/ActionSummary"
          },
          "log": {
            "$ref": "#/components/schemas/ActionLogSummary"
          }
        },
        "type": "object",
        "required": [
          "action",
          "log"
        ],
        "title": "ActionExecuted"
      },
      "ActionKind": {
        "type": "string",
        "enum": [
          "tag_incident",
          "elevate_severity",
          "flag_host_in_lab",
          "quarantine_host_lab",
          "invalidate_lab_session",
          "block_observable",
          "kill_process_lab",
          "request_evidence"
        ],
        "title": "ActionKind"
      },
      "ActionLogSummary": {
        "properties": {
          "executed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Executed At"
          },
          "executed_by": {
            "type": "string",
            "title": "Executed By"
          },
          "result": {
            "type": "string",
            "enum": [
              "ok",
              "fail",
              "skipped"
            ],
            "title": "Result"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "executed_at",
          "executed_by",
          "result",
          "reason"
        ],
        "title": "ActionLogSummary"
      },
      "ActionProposeIn": {
        "properties": {
          "incident_id": {
            "type": "string",
            "format": "uuid",
            "title": "Incident Id"
          },
          "kind": {
            "$ref": "#/components/schemas/ActionKind"
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params"
          }
        },
        "type": "object",
        "required": [
          "incident_id",
          "kind",
          "params"
        ],
        "title": "ActionProposeIn"
      },
      "ActionProposed": {
        "properties": {
          "action": {
            "$ref": "#/components/schemas/ActionSummary"
          }
        },
        "type": "object",
        "required": [
          "action"
        ],
        "title": "ActionProposed"
      },
      "ActionSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "classification": {
            "type": "string",
            "enum": [
              "auto_safe",
              "suggest_only",
              "reversible",
              "disruptive"
            ],
            "title": "Classification"
          },
          "classification_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Classification Reason"
          },
          "status": {
            "type": "string",
            "enum": [
              "proposed",
              "executed",
              "failed",
              "skipped",
              "reverted"
            ],
            "title": "Status"
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params"
          },
          "proposed_by": {
            "type": "string",
            "enum": [
              "system",
              "analyst"
            ],
            "title": "Proposed By"
          },
          "proposed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Proposed At"
          },
          "last_log": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActionLogSummary"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "classification",
          "classification_reason",
          "status",
          "params",
          "proposed_by",
          "proposed_at",
          "last_log"
        ],
        "title": "ActionSummary"
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LabAssetIn": {
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/LabAssetKind"
          },
          "natural_key": {
            "type": "string",
            "title": "Natural Key"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "natural_key"
        ],
        "title": "LabAssetIn"
      },
      "LabAssetKind": {
        "type": "string",
        "enum": [
          "user",
          "host",
          "ip",
          "observable"
        ],
        "title": "LabAssetKind"
      },
      "LabAssetOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "natural_key": {
            "type": "string",
            "title": "Natural Key"
          },
          "registered_at": {
            "type": "string",
            "format": "date-time",
            "title": "Registered At"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "natural_key",
          "registered_at",
          "notes"
        ],
        "title": "LabAssetOut"
      },
...
      "ResponseList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ActionSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "ResponseList"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}